use uintptr_t instead of int for debug printf of pointers

+++ src/rfcomm.c
@@ -809,7 +809,7 @@ static void rfcomm_multiplexer_state_machine(rfcomm_multiplexer_t * multiplexer,
         case RFCOMM_MULTIPLEXER_SEND_SABM_0:
             switch (event) {
                 case MULT_EV_READY_TO_SEND:
-                    log_info("Sending SABM #0 - (multi 0x%08x)\n", (int) multiplexer);
+                    log_info("Sending SABM #0 - (multi 0x%08x)\n", (uintptr_t) multiplexer);
                     multiplexer->state = RFCOMM_MULTIPLEXER_W4_UA_0;
                     rfcomm_send_sabm(multiplexer, 0);
                     break;
@@ -1360,7 +1360,7 @@ static void rfcomm_channel_state_machine(rfcomm_channel_t *channel, rfcomm_chann
         case RFCOMM_CHANNEL_SEND_UIH_PN:
             switch (event->type) {
                 case CH_EVT_READY_TO_SEND:
-                    log_info("Sending UIH Parameter Negotiation Command for #%u (channel 0x%08x\n", channel->dlci, (int) channel );
+                    log_info("Sending UIH Parameter Negotiation Command for #%u (channel 0x%08x\n", channel->dlci, (uintptr_t) channel );
                     channel->state = RFCOMM_CHANNEL_W4_PN_RSP;
                     rfcomm_send_uih_pn_command(multiplexer, channel->dlci, channel->max_frame_size);
                     break;
